move testing files near what they test
