+++ frontend/src/data/pairing_test.js
@@ -1,21 +1,9 @@
 import React from 'react';
 import { renderHook, act } from '@testing-library/react-hooks'
-import { useParticipants } from '../../src/data/participants';
-import { usePairs, usePostPairing } from '../../src/data/pairing';
-import { api } from '../helper';
+import { usePairs, usePostPairing } from './pairing';
+import { api } from '../../test/helper';
 
 jest.mock('axios-hooks');
-
-describe('useParticipants', () => {
-  // NOTE: testing here is light since this just calls through to axios
-  test('gets some participants', async () => {
-    api.mock(api.participants({id: 1}));
-    const { result } = renderHook(() => useParticipants());
-    expect(result.current[0].loading).toBeFalsy();
-    expect(result.current[0].error).toBeFalsy();
-    expect(result.current[0].data).toStrictEqual([{id: 1}]);
-  });
-});
 
 describe('usePairs', () => {
   // NOTE: testing here is light since this just calls through to axios
@@ -46,3 +34,4 @@ describe('usePostPairing', () => {
     expect(posted).toStrictEqual({ data: { mentor: 1, learner: 2 } });
   });
 });
+
